--- app/src/main/java/com/homies/homies/tasks/TaskDetail.java
@@ -1,4 +1,0 @@
-package com.homies.homies.tasks;
-
-public class TaskDetail {
-}
