--- app/src/main/java/com/homies/homies/tasks/TaskEdit.java
@@ -1,4 +1,0 @@
-package com.homies.homies.tasks;
-
-public class TaskEdit {
-}
