+++ app/src/main/res/layout/activity_splash.xml
@@ -8,7 +8,7 @@
     android:background="@color/primaryColor">
 
     <ImageView
-        android:id="@+id/imageView"
+        android:id="@+id/logo"
         android:layout_width="133dp"
         android:layout_height="164dp"
         android:layout_marginStart="139dp"
@@ -22,7 +22,7 @@
         app:srcCompat="@drawable/ic_logo_circle" />
 
     <ImageView
-        android:id="@+id/imageView2"
+        android:id="@+id/nameLogo"
         android:layout_width="146dp"
         android:layout_height="69dp"
         android:layout_marginStart="132dp"
@@ -41,7 +41,7 @@
         android:layout_marginStart="181dp"
         android:layout_marginEnd="182dp"
         android:layout_marginBottom="38dp"
-        app:layout_constraintBottom_toTopOf="@+id/imageView2"
+        app:layout_constraintBottom_toTopOf="@+id/nameLogo"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         android:indeterminate="true"
