+++ app/src/main/res/layout/activity_user.xml
@@ -8,7 +8,7 @@
     android:id="@+id/main_layout">
 
     <ImageView
-        android:id="@+id/logoIV"
+        android:id="@+id/logo"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="51dp"
@@ -27,7 +27,7 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/logoIV" />
+        app:layout_constraintTop_toBottomOf="@+id/logo" />
