+++ app/src/main/res/layout/dialog_add_user.xml
@@ -51,7 +51,7 @@
             android:drawablePadding="10dp"
             android:drawableTint="@color/primaryColor"
             android:fontFamily="@font/montserrat_regular"
-            android:hint="@string/add_user_et"
+            android:hint="@string/username"
             android:inputType="text"
             android:maxLength="50"
             android:padding="15dp"
@@ -70,7 +70,7 @@
             android:background="@drawable/btn_custom"
             android:fontFamily="@font/montserrat_regular"
             android:padding="1dp"
-            android:text="@string/confirmUser"
+            android:text="@string/confirm"
             android:textAllCaps="false"
             android:textColor="@color/primaryTextColor"
             app:layout_constraintBottom_toBottomOf="parent"
