+++ app/src/main/res/layout/dialog_delete_profile.xml
