+++ app/src/main/res/layout/dialog_forget_pass.xml
@@ -36,7 +36,7 @@
             android:layout_marginEnd="20dp"
             android:background="@drawable/rounded_edittext"
             android:fontFamily="@font/montserrat_regular"
-            android:hint="@string/recoverInput"
+            android:hint="@string/e_mail"
             android:inputType="textEmailAddress"
             android:maxLength="50"
             android:padding="15dp"
