+++ app/src/main/res/layout/fragment_lists_group.xml
@@ -6,7 +6,6 @@
 
 
     <androidx.constraintlayout.widget.ConstraintLayout
-
         android:layout_width="match_parent"
         android:layout_height="match_parent">
