+++ app/src/main/res/layout/fragment_settings_user.xml
@@ -24,7 +24,7 @@
             android:layout_gravity="center"
             android:layout_marginTop="8dp"
             android:layout_marginBottom="5dp"
-            android:src="@drawable/rounded_photo"
+            android:src="@drawable/user_profile_icon"
             app:civ_border_color="@color/secondaryColor"
             app:civ_border_width="2dp"
             app:layout_constraintBottom_toTopOf="@+id/ip_user"
