+++ app/src/main/res/layout/fragment_shopping_list.xml
@@ -10,7 +10,7 @@
         android:id="@+id/haveTo"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="@string/por_comprar"
+        android:text="@string/to_buy"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
@@ -27,7 +27,7 @@
         android:id="@+id/comprado"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="@string/comprado"
+        android:text="@string/purchased"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/porcomprar" />
@@ -46,7 +46,7 @@
         android:layout_height="wrap_content"
         android:layout_marginTop="20dp"
         app:layout_constraintEnd_toEndOf="parent"
-        android:text="@string/borrar_todo"
+        android:text="@string/delete"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/yacomprado" />
